requirements.txt
# Main requirements for cinema-graph project
fastapi==0.115.0
uvicorn[standard]==0.30.0
neo4j==5.25.0
google-generativeai==0.8.0
pydantic==2.9.0
python-dotenv==1.0.1
requests
